+++ orchestrator/ethereum.go
@@ -1,0 +1,39 @@
+package orchestrator
+
+import (
+	"context"
+	"math/big"
+
+	eth "github.com/ethereum/go-ethereum/common"
+	ethtypes "github.com/ethereum/go-ethereum/core/types"
+
+	peggyevents "github.com/InjectiveLabs/peggo/solidity/wrappers/Peggy.sol"
+	peggytypes "github.com/InjectiveLabs/sdk-go/chain/peggy/types"
+)
+
+// EthereumNetwork is the orchestrator's reference endpoint to the Ethereum network
+type EthereumNetwork interface {
+	FromAddress() eth.Address
+	HeaderByNumber(ctx context.Context, number *big.Int) (*ethtypes.Header, error)
+	GetPeggyID(ctx context.Context) (eth.Hash, error)
+
+	GetSendToCosmosEvents(startBlock, endBlock uint64) ([]*peggyevents.PeggySendToCosmosEvent, error)
+	GetSendToInjectiveEvents(startBlock, endBlock uint64) ([]*peggyevents.PeggySendToInjectiveEvent, error)
+	GetPeggyERC20DeployedEvents(startBlock, endBlock uint64) ([]*peggyevents.PeggyERC20DeployedEvent, error)
+	GetValsetUpdatedEvents(startBlock, endBlock uint64) ([]*peggyevents.PeggyValsetUpdatedEvent, error)
+	GetTransactionBatchExecutedEvents(startBlock, endBlock uint64) ([]*peggyevents.PeggyTransactionBatchExecutedEvent, error)
+
+	GetValsetNonce(ctx context.Context) (*big.Int, error)
+	SendEthValsetUpdate(ctx context.Context,
+		oldValset *peggytypes.Valset,
+		newValset *peggytypes.Valset,
+		confirms []*peggytypes.MsgValsetConfirm,
+	) (*eth.Hash, error)
+
+	GetTxBatchNonce(ctx context.Context, erc20ContractAddress eth.Address) (*big.Int, error)
+	SendTransactionBatch(ctx context.Context,
+		currentValset *peggytypes.Valset,
+		batch *peggytypes.OutgoingTxBatch,
+		confirms []*peggytypes.MsgConfirmBatch,
+	) (*eth.Hash, error)
+}
