+++ orchestrator/orchestrator.go
@@ -2,89 +2,15 @@ package orchestrator
 
 import (
 	"context"
-	"math/big"
 	"time"
 
 	eth "github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/pkg/errors"
 	log "github.com/xlab/suplog"
 
 	"github.com/InjectiveLabs/metrics"
 	"github.com/InjectiveLabs/peggo/orchestrator/loops"
-	peggyevents "github.com/InjectiveLabs/peggo/solidity/wrappers/Peggy.sol"
-	peggytypes "github.com/InjectiveLabs/sdk-go/chain/peggy/types"
 )
-
-type PriceFeed interface {
-	QueryUSDPrice(address eth.Address) (float64, error)
-}
-
-type InjectiveNetwork interface {
-	PeggyParams(ctx context.Context) (*peggytypes.Params, error)
-	GetBlockCreationTime(ctx context.Context, height int64) (time.Time, error)
-
-	// claims
-	LastClaimEvent(ctx context.Context) (*peggytypes.LastClaimEvent, error)
-	SendEthereumClaims(
-		ctx context.Context,
-		lastClaimEvent uint64,
-		oldDeposits []*peggyevents.PeggySendToCosmosEvent,
-		deposits []*peggyevents.PeggySendToInjectiveEvent,
-		withdraws []*peggyevents.PeggyTransactionBatchExecutedEvent,
-		erc20Deployed []*peggyevents.PeggyERC20DeployedEvent,
-		valsetUpdates []*peggyevents.PeggyValsetUpdatedEvent,
-	) error
-
-	// batches
-	UnbatchedTokenFees(ctx context.Context) ([]*peggytypes.BatchFees, error)
-	SendRequestBatch(ctx context.Context, denom string) error
-	OldestUnsignedTransactionBatch(ctx context.Context) (*peggytypes.OutgoingTxBatch, error)
-	SendBatchConfirm(ctx context.Context, peggyID eth.Hash, batch *peggytypes.OutgoingTxBatch, ethFrom eth.Address) error
-	LatestTransactionBatches(ctx context.Context) ([]*peggytypes.OutgoingTxBatch, error)
-	TransactionBatchSignatures(ctx context.Context, nonce uint64, tokenContract eth.Address) ([]*peggytypes.MsgConfirmBatch, error)
-
-	// valsets
-	OldestUnsignedValsets(ctx context.Context) ([]*peggytypes.Valset, error)
-	SendValsetConfirm(ctx context.Context, peggyID eth.Hash, valset *peggytypes.Valset, ethFrom eth.Address) error
-	LatestValsets(ctx context.Context) ([]*peggytypes.Valset, error)
-	AllValsetConfirms(ctx context.Context, nonce uint64) ([]*peggytypes.MsgValsetConfirm, error)
-	ValsetAt(ctx context.Context, nonce uint64) (*peggytypes.Valset, error)
-}
-
-type EthereumNetwork interface {
-	FromAddress() eth.Address
-	HeaderByNumber(ctx context.Context, number *big.Int) (*types.Header, error)
-	GetPeggyID(ctx context.Context) (eth.Hash, error)
-
-	// events
-	GetSendToCosmosEvents(startBlock, endBlock uint64) ([]*peggyevents.PeggySendToCosmosEvent, error)
-	GetSendToInjectiveEvents(startBlock, endBlock uint64) ([]*peggyevents.PeggySendToInjectiveEvent, error)
-	GetPeggyERC20DeployedEvents(startBlock, endBlock uint64) ([]*peggyevents.PeggyERC20DeployedEvent, error)
-	GetValsetUpdatedEvents(startBlock, endBlock uint64) ([]*peggyevents.PeggyValsetUpdatedEvent, error)
-	GetTransactionBatchExecutedEvents(startBlock, endBlock uint64) ([]*peggyevents.PeggyTransactionBatchExecutedEvent, error)
-
-	// valsets
-	GetValsetNonce(ctx context.Context) (*big.Int, error)
-	SendEthValsetUpdate(
-		ctx context.Context,
-		oldValset *peggytypes.Valset,
-		newValset *peggytypes.Valset,
-		confirms []*peggytypes.MsgValsetConfirm,
-	) (*eth.Hash, error)
-
-	// batches
-	GetTxBatchNonce(
-		ctx context.Context,
-		erc20ContractAddress eth.Address,
-	) (*big.Int, error)
-	SendTransactionBatch(
-		ctx context.Context,
-		currentValset *peggytypes.Valset,
-		batch *peggytypes.OutgoingTxBatch,
-		confirms []*peggytypes.MsgConfirmBatch,
-	) (*eth.Hash, error)
-}
 
 const defaultLoopDur = 60 * time.Second
 
@@ -92,8 +18,8 @@ type PeggyOrchestrator struct {
 	logger  log.Logger
 	svcTags metrics.Tags
 
-	injective InjectiveNetwork
-	ethereum  EthereumNetwork
+	inj       InjectiveNetwork
+	eth       EthereumNetwork
 	pricefeed PriceFeed
 
 	erc20ContractMapping map[eth.Address]string
@@ -118,11 +44,11 @@ func NewPeggyOrchestrator(
 	valsetRelayingOffset,
 	batchRelayingOffset string,
 ) (*PeggyOrchestrator, error) {
-	orch := &PeggyOrchestrator{
+	o := &PeggyOrchestrator{
 		logger:               log.DefaultLogger,
 		svcTags:              metrics.Tags{"svc": "peggy_orchestrator"},
-		injective:            injective,
-		ethereum:             ethereum,
+		inj:                  injective,
+		eth:                  ethereum,
 		pricefeed:            priceFeed,
 		erc20ContractMapping: erc20ContractMapping,
 		minBatchFeeUSD:       minBatchFeeUSD,
@@ -137,7 +63,7 @@ func NewPeggyOrchestrator(
 			return nil, errors.Wrapf(err, "valset relaying enabled but offset duration is not properly set")
 		}
 
-		orch.relayValsetOffsetDur = dur
+		o.relayValsetOffsetDur = dur
 	}
 
 	if batchRelayingEnabled {
@@ -146,10 +72,10 @@ func NewPeggyOrchestrator(
 			return nil, errors.Wrapf(err, "batch relaying enabled but offset duration is not properly set")
 		}
 
-		orch.relayBatchOffsetDur = dur
+		o.relayBatchOffsetDur = dur
 	}
 
-	return orch, nil
+	return o, nil
 }
 
 // Run starts all major loops required to make
